Page Cookie
Set a timeout function which wouldn't remove with cleartimeout function
so got it to check for a page instead. Timeout function causes cardbacks
from incomplete timeout functions to appear on the menu page. Therefore
the timeout function now checks a cookie that tells it that it is not on
the menupage.